Added test for vsock after snapshot
